% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{config_from_yaml}
\alias{config_from_yaml}
\title{Load configuration from YAML file}
\usage{
config_from_yaml(path)
}
\arguments{
\item{path}{character path to YAML config file}
}
\value{
list of merged configuration parameters
}
\description{
Load a YAML config and merge with defaults from triage_rules()
}
