% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{export_sem_report}
\alias{export_sem_report}
\title{Export a compact HTML report with pre/post fit and action log}
\usage{
export_sem_report(dat, model, res, file = "sem_report.html")
}
\arguments{
\item{dat}{data}

\item{model}{model syntax}

\item{res}{result from triage_apply}

\item{file}{output .html file path}
}
\value{
normalized file path (invisible)
}
\description{
If gt is available, includes an APA-like fit table. Otherwise plain HTML.
}
