% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lower_order_equivalent}
\alias{lower_order_equivalent}
\title{Create lower-order equivalent model}
\usage{
lower_order_equivalent(model, dat = NULL)
}
\arguments{
\item{model}{character lavaan model syntax for higher-order model}

\item{dat}{data.frame (optional, for validation)}
}
\value{
character lavaan syntax for lower-order equivalent model
}
\description{
Convert higher-order model to lower-order equivalent by removing higher-order factors
}
